// mobile/src/lib/types.ts
export type Expense = {
  id: string;
  user_id: string;
  category_id: string | null;
  vendor_id: string | null;
  description: string;
  amount: number;
  paid: boolean;
  due_date: string | null;
  created_at: string;
}